[iOS] refactor :recycle: LoadingUI 구현부분 분리
#132
