+++ iOS/DoCollabo/DoCollabo/Issues/Controllers/IssuesViewController.swift
@@ -150,51 +150,6 @@ extension IssuesViewController: HeaderViewActionDelegate {
     }
 }
 
-// MARK:- Loading UI
-
-extension IssuesViewController {
-    private func hideViews() {
-        titleHeaderBackgroundView.alpha = 0
-        titleHeaderView.alpha = 0
-        issuesCollectionView.alpha = 0
-        activityIndicator.alpha = 1
-        activityIndicator.startAnimating()
-    }
-    
-    private func showViews() {
-        UIView.animate(
-            withDuration: 0.7,
-            delay: 0,
-            usingSpringWithDamping: 1,
-            initialSpringVelocity: 1,
-            options: .curveEaseOut,
-            animations: {
-                self.titleHeaderBackgroundView.alpha = 1
-                self.activityIndicator.alpha = 0
-        }, completion: { _ in
-            self.activityIndicator.stopAnimating()
-        })
-        UIView.animate(
-            withDuration: 1,
-            delay: 0.2,
-            usingSpringWithDamping: 1,
-            initialSpringVelocity: 1,
-            options: .curveEaseOut,
-            animations: {
-                self.titleHeaderView.alpha = 1
-        })
-        UIView.animate(
-            withDuration: 0.5,
-            delay: 0.4,
-            usingSpringWithDamping: 1.1,
-            initialSpringVelocity: 1,
-            options: .curveEaseOut,
-            animations: {
-                self.issuesCollectionView.alpha = 1
-        })
-    }
-}
-
 // MARK:- Configuration
 
 extension IssuesViewController {
